use imported varint library

+++ packages/ethereum-contracts/contracts/CID_External.sol
@@ -7,7 +7,7 @@ contract CID_External {
   function createDigest(uint64 code, bytes memory digest) public pure returns (uint256, bytes memory) {
     return CID.createDigest(code, digest);
   }
-  function encodeCID(uint256 version, uint256 code, bytes memory multihash) public pure returns (bytes memory)  {
+  function encodeCID(uint64 version, uint64 code, bytes memory multihash) public pure returns (bytes memory)  {
     return CID.encodeCID(version, code, multihash);
   }
 }
